pyproject.toml
[project]
name = "playlist-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
]
